+++ strelyaev.roman/P4/clock_matrix.cpp
@@ -1,0 +1,33 @@
+#include "clock_matrix.h"
+void clockwise(int * a, int rows, int columns)
+{
+  int num = 1;
+  for (int delta = 0; delta < std::max(rows - 2, 1); delta++)
+  {
+    for (int i = (rows-1)*columns - delta*(columns-1); i > delta*(columns+1); i-=columns)
+    {
+      a[i] = num++;
+    }
+    for (int i  = 0 + delta*(columns+1); i < columns + delta*(columns-1); i+=1)
+    {
+      a[i] = num++;
+    }
+    --num;
+    for (int i  = (columns-1) + delta*(columns-1); i < rows*columns - delta*(columns+1) - 1; i+=columns)
+    {
+      a[i] = num++;
+    }
+    for (int i  = rows*columns - 1 - delta*(columns+1); i > (rows-1) * columns - delta*(columns-1); i-=1)
+    {
+      a[i] = num++;
+    }
+  }
+}
+
+void matrixSubtraction(int * original, int * reverse, int n)
+{
+  for (int i = 0; i < n; i++)
+  {
+    original[i] -= reverse[i];
+  }
+}
