+++ strelyaev.roman/P4/clock_matrix.h
@@ -1,0 +1,6 @@
+#ifndef CLOCK_MATRIX_H
+#define CLOCK_MATRIX_H
+#include <algorithm>
+void clockwise(int * a, int rows, int columns);
+void matrixSubtraction(int * original, int * reverse, int n);
+#endif
